+++ src/dom/special_properties.ts
@@ -1,0 +1,11 @@
+export interface ValueableHTMLElement extends HTMLElement {
+  [SpecialProperty.Value]: string;
+}
+
+/**
+ * Properties that must be updated by library
+ * in order to show reactivity.
+ */
+export enum SpecialProperty {
+  Value = 'value',
+}