+++ test/template.test.ts
@@ -64,7 +64,7 @@ describe('Test template of "edelweiss"', () => {
   });
 
   test('add updated hook to element that has value attribute', async () => {
-    const input = await html` <input value=${() => 'Hoooo'} /> `;
+    const input = await html` <input .value=${() => 'Hoooo'} /> `;
 
     expect(input).toMatch(/value/);
     expect(input).toMatch(/data-hook-id-updated/);